+++ cpp/cmake/dnvVistaSdkCppDependencies.cmake
@@ -21,6 +21,7 @@ set(CMAKE_FIND_QUIETLY      ON     )
 set(VISTA_SDK_CPP_ZLIB_NG_VERSION                "2.3.1")
 set(VISTA_SDK_CPP_DEPS_NFX_SERIALIZATION_VERSION "0.3.0")
 set(VISTA_SDK_CPP_DEPS_NFX_STRINGBUILDER_VERSION "0.4.1")
+set(VISTA_SDK_CPP_DEPS_NFX_STRINGUTILS_VERSION   "0.3.0")
 
 #----------------------------------------------
 # FetchContent dependencies
@@ -129,6 +130,32 @@ endif()
 
 if(NOT nfx-stringbuilder_FOUND)
 	FetchContent_MakeAvailable(nfx-stringbuilder)
+endif()
+
+# --- nfx-stringutils ---
+find_package(nfx-stringutils ${VISTA_SDK_CPP_DEPS_NFX_STRINGUTILS_VERSION} EXACT QUIET)
+if(NOT nfx-stringutils_FOUND)
+	set(NFX_STRINGUTILS_BUILD_TESTS         OFF CACHE BOOL "Build tests"                 FORCE)
+	set(NFX_STRINGUTILS_BUILD_SAMPLES       OFF CACHE BOOL "Build samples"               FORCE)
+	set(NFX_STRINGUTILS_BUILD_BENCHMARKS    OFF CACHE BOOL "Build benchmarks"            FORCE)
+	set(NFX_STRINGUTILS_BUILD_DOCUMENTATION OFF CACHE BOOL "Build Doxygen documentation" FORCE)
+	set(NFX_STRINGUTILS_INSTALL_PROJECT     OFF CACHE BOOL "Install project"             FORCE)
+	set(NFX_STRINGUTILS_PACKAGE_SOURCE      OFF CACHE BOOL "Enable source package"       FORCE)
+	set(NFX_STRINGUTILS_PACKAGE_ARCHIVE     OFF CACHE BOOL "Enable archive package"      FORCE)
+	set(NFX_STRINGUTILS_PACKAGE_DEB         OFF CACHE BOOL "Enable DEB package"          FORCE)
+	set(NFX_STRINGUTILS_PACKAGE_RPM         OFF CACHE BOOL "Enable RPM package"          FORCE)
+	set(NFX_STRINGUTILS_PACKAGE_WIX         OFF CACHE BOOL "Enable WIX package"          FORCE)
+
+	FetchContent_Declare(
+		nfx-stringutils
+		GIT_REPOSITORY https://github.com/nfx-libs/nfx-stringutils
+		GIT_TAG        ${VISTA_SDK_CPP_DEPS_NFX_STRINGUTILS_VERSION}
+		GIT_SHALLOW    TRUE
+	)
+endif()
+
+if(NOT nfx-stringutils_FOUND)
+	FetchContent_MakeAvailable(nfx-stringutils)
 endif()
 
 #----------------------------------------------