+++ cpp/include/dnv/vista/sdk/Locations.h
@@ -69,6 +69,7 @@ namespace dnv::vista::sdk
 	class Location final
 	{
 		friend class Locations;
+		friend class LocationBuilder;
 
 	private:
 		/**
@@ -231,6 +232,7 @@ namespace dnv::vista::sdk
 	class Locations final
 	{
 		friend class VIS;
+		friend class LocationBuilder;
 
 	private:
 		/**